- Upgrade to 2.4.1 client: thanks to Dereka, Vladimir, Neo2003 and me.   [New: 5651_characters_characters.sql] Update for `character`.`data` column to 2.4.1 version.   [New: 5651_characters_character_social.sql] Update for character_social table (new column `note` added).   [New: 5651_characters_guild_bank_tab.sql] Update for guild_bank_tab table (new column `TabText` added).   [New: 5651_mangos_creature_template.sql]] Update for creature_template table.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5651 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Creature.h
@@ -269,7 +269,7 @@ struct CreatureModelInfo
     uint32 modelid;
     float bounding_radius;
     float combat_reach;
-    uint32 gender;
+    uint8 gender;
     uint32 modelid_other_gender;
 };
