+++ trunk/src/game/GossipDef.h
@@ -180,7 +180,7 @@ class MANGOS_DLL_SPEC PlayerMenu
         /*********************************************************/
         /***                    QUEST SYSTEM                   ***/
         /*********************************************************/
-        void SendQuestGiverStatus( uint32 questStatus, uint64 npcGUID );
+        void SendQuestGiverStatus( uint8 questStatus, uint64 npcGUID );
 
         void SendQuestGiverQuestList( QEmote eEmote, std::string Title, uint64 npcGUID );
